added script to convert course api data to pickled, formatted data for reading
